chore: dead code cleanup and square rename
