+++ apps/web/app/api/v1/integrations/square/test/route.ts
@@ -5,7 +5,7 @@ import { logger } from "@/lib/logger";
 import {
   loadSquareSettings,
   testSquareConnection,
-} from "@/lib/integrations/square";
+} from "@/lib/integrations/square-payments";
 
 export const dynamic = "force-dynamic";
