--- apps/web/lib/email/templates.ts
@@ -1,21 +1,0 @@
-export {
-  type InvoiceEmailData,
-  type EstimateEmailData,
-  type VisitReminderEmailData,
-  type InvoiceFollowupEmailData,
-  type OnMyWayEmailData,
-  type BookingConfirmedEmailData,
-  type IntakeInviteEmailData,
-  type ReviewRequestEmailData,
-  invoiceEmailHtml,
-  invoiceEmailText,
-  estimateEmailHtml,
-  estimateEmailText,
-  visitReminderEmailHtml,
-  invoiceFollowupEmailHtml,
-  onMyWayEmailHtml,
-  bookingConfirmedEmailHtml,
-  intakeInviteEmailHtml,
-  intakeInviteEmailText,
-  reviewRequestEmailHtml,
-} from "@ai-fsm/email-templates";
